Rename sendLogs.ts to logging.ts
